Update package for level sets
