+++ src/FastIce.jl
@@ -1,5 +1,6 @@
 module FastIce
 
 include("level_sets/level_sets.jl")
+include("geometry.jl")
 
 end # module
